pyproject.toml
[project]
name = "miacro"
version = "0.1.0"
description = "Community Intelligence Service"
authors = [{ name = "MicroClub-USTHB" }]

readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.1",
    "pydantic-settings>=2.13.1",
    "uvicorn[standard]>=0.41.0",
    "qdrant-client>=1.7.0",
    "pytest>=9.0.2",
    "google-genai>=1.66.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "prometheus-client>=0.24.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-asyncio>=1.3.0",
]
